+++ packages/algob/test/internal/cli/project-creation.ts
@@ -18,66 +18,71 @@ describe("Init project", () => {
 		}
 	});
 
-	it("should init project in an empty folder(javascript) with infrastructure folder", async () => {
-		const location = "test-project";
-		await createProject(location, false, true);
-
+	function checkPaths(
+		location: string,
+		isTs: boolean,
+		withInfrastructure: boolean,
+		isNpm: boolean
+	) {
 		assert.isTrue(fs.existsSync(`./${location}/algob.config.js`));
-		assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.js`));
-		assert.isTrue(fs.existsSync(`./${location}/infrastructure`));
+		if (isTs) {
+			assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.ts`));
+		} else {
+			assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.js`));
+		}
+		assert.equal(withInfrastructure, fs.existsSync(`./${location}/infrastructure`));
+	}
+
+	it("should init npm project in an empty folder(javascript) with infrastructure folder", async () => {
+		const location = "test-project";
+		await createProject(location, false, true, true);
+		checkPaths(location, false, true, true);
 	});
 
-	it("should init project in an empty folder(javascript) without infrastructure folder", async () => {
+	it("should init npm project in an empty folder(javascript) without infrastructure folder", async () => {
 		const location = "test-project";
-		await createProject(location, false, false);
-
-		assert.isTrue(fs.existsSync(`./${location}/algob.config.js`));
-		assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.js`));
-		assert.isFalse(fs.existsSync(`./${location}/infrastructure`));
+		await createProject(location, false, false, true);
+		checkPaths(location, false, false, true);
 	});
 
-	it("should init project in a empty folder(typescript) with infrastructure folder", async () => {
+	it("should init npm project in a empty folder(typescript) with infrastructure folder", async () => {
 		const location = "test-project";
-		await createProject(location, true, true);
-
-		assert.isTrue(fs.existsSync(`./${location}/algob.config.js`));
-		assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.ts`));
-		assert.isTrue(fs.existsSync(`./${location}/tsconfig.json`));
-		assert.isTrue(fs.existsSync(`./${location}/infrastructure/Makefile`));
+		await createProject(location, true, true, true);
+		checkPaths(location, true, true, true);
 	});
 
-	it("should init project in a empty folder(typescript) without infrastructure folder", async () => {
+	it("should init npm project in a empty folder(typescript) without infrastructure folder", async () => {
 		const location = "test-project";
-		await createProject(location, true, false);
-
-		assert.isTrue(fs.existsSync(`./${location}/algob.config.js`));
-		assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.ts`));
-		assert.isTrue(fs.existsSync(`./${location}/tsconfig.json`));
-		assert.isFalse(fs.existsSync(`./${location}/infrastructure/Makefile`));
+		await createProject(location, true, false, true);
+		checkPaths(location, true, false, true);
 	});
 
-	it("should not create project if folder already exist", async () => {
-		await createProject("location", false, false);
+	it("should init yarn project in a empty folder(typescript) without infrastructure folder", async () => {
+		const location = "test-project";
+		await createProject(location, true, false, false);
+		checkPaths(location, true, false, true);
+	});
+
+	it("should not create an npm project if folder already exist", async () => {
+		await createProject("location", false, false, true);
 
 		await expectBuilderErrorAsync(
-			async () => await createProject("location", false, false),
+			async () => await createProject("location", false, false, true),
 			ERRORS.GENERAL.INIT_INSIDE_PROJECT
 		);
 	});
 
-	it("should init project in an empty folder(typescript) with `.`", async () => {
+	it("should init npm project in an empty folder(typescript) with `.`", async () => {
 		const location = ".";
-		await createProject(location, true, false);
-
-		assert.isTrue(fs.existsSync(`./${location}/algob.config.js`));
-		assert.isTrue(fs.existsSync(`./${location}/scripts/0-sampleScript.ts`));
+		await createProject(location, true, false, true);
+		checkPaths(location, true, false, true);
 	});
 
-	it("should not init project if it already exists with `.`", async () => {
-		await createProject(".", false, false);
+	it("should not init npm project if it already exists with `.`", async () => {
+		await createProject(".", false, false, true);
 
 		await expectBuilderErrorAsync(
-			async () => await createProject(".", false, false),
+			async () => await createProject(".", false, false, true),
 			ERRORS.GENERAL.INIT_INSIDE_PROJECT
 		);
 	});