+++ CHANGELOG.md
@@ -14,9 +14,18 @@ Features, Bug Fixes, API Breaking, Deprecated, Infrastructure, Template Updates
 
 ## Unreleased
 
-### Bug Fixes 
+- Updated yarn to v3.2.1
+- Changed default sample project license to ISC
 
-- Fix `txn AssetSender` should return zero address by default. 
+### Features
+
+- Add `--npm` flag to `algob init` and `algob unbox`. Note: by default we will use `yarn`.
+- Improved `algob/project-dev-script.sh` which is script setting up a local project.
+
+### Bug Fixes
+
+- Fix `txn AssetSender` should return zero address by default.
+- Fix `KMDCredentialsFromEnv` loading using KMD_DATA. Algob was trying to use `env.$KMD_DATA` instead of `env.KMD_DATA`
 
 ## v5.0.1 2022-07-11